+++ frontend/src/schemas/protocols/inbound/shadowsocks.ts
@@ -1,15 +1,6 @@
 import { z } from 'zod';
 
-export const SSMethodSchema = z.enum([
-  'aes-256-gcm',
-  'chacha20-poly1305',
-  'chacha20-ietf-poly1305',
-  'xchacha20-ietf-poly1305',
-  '2022-blake3-aes-128-gcm',
-  '2022-blake3-aes-256-gcm',
-  '2022-blake3-chacha20-poly1305',
-]);
-export type SSMethod = z.infer<typeof SSMethodSchema>;
+import { SSMethodSchema } from '../shared/shadowsocks';
 
 export const SSNetworkSchema = z.enum(['tcp', 'udp', 'tcp,udp']);
 export type SSNetwork = z.infer<typeof SSNetworkSchema>;
