+++ frontend/src/schemas/protocols/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './shadowsocks';
+export * from './vmess';
